requirements/development.txt
-r ./base.txt

django-nose
Pygments
sphinx
